rename curatedEventList to curatedEvents
