+++ src/domain/collection/curatedEvents/EventCards.tsx
