+++ src/domain/collection/curatedEvents/OnlyExpiredEvents.tsx
