+++ src/domain/collection/curatedEvents/__tests__/CuratedEvents.test.tsx
