+++ src/domain/collection/curatedEvents/__tests__/EventCards.test.tsx
