+++ src/domain/collection/curatedEvents/__tests__/OnlyExpiredEvents.test.tsx
