+++ src/domain/collection/curatedEvents/curatedEvents.module.scss
