+++ src/domain/collection/curatedEvents/eventCards.module.scss
