+++ src/domain/collection/curatedEvents/utils.tsx
